# lib/api-zod/src/generated/types/hostelRoom.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student ERP System API
 * OpenAPI spec version: 0.1.0
 */
import type { HostelRoomType } from "./hostelRoomType";

export interface HostelRoom {
  id: number;
  roomNumber: string;
  block: string;
  floor: number;
  capacity: number;
  occupied: number;
  type: HostelRoomType;
  amenities: string[];
  available: boolean;
}