# lib/api-zod/src/generated/types/updateStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStudentBody {
  phone?: string;
  address?: string;
  semester?: string;
}